Replace old by new Card implementation
