+++ docs/model_card.rst
@@ -42,32 +42,30 @@ to touch it yourself.
 
 The markdown part does not necessarily need to follow any specification in
 terms of information passed, which gives the user a lot of flexibility. The
-markdown part of the ``README.md`` file is generated from a Jinja template
-with slots that you can inject your content in. ``skops`` has a default
-template which includes the following slots for free text sections:
+markdown part of the ``README.md`` file comes with a couple of defaults provided
+by ``skops``, which includes the following slots for free text sections:
 
-- ``"model_description"``: A description of the model.
-- ``"limitations"``: Intended use for the model, limitations and potential
-  biases. This section should also include risks of using models in certain
-  domains if relevant.
-- ``"get_started_code"``: Code the user can run to load and use the model.
-- ``"model_card_authors"``: Authors of the model card. This section includes
-  authors of the model card, while ``"citation_bibtex"`` includes citations
-  related to the model if relevant.
-- ``"model_card_contact"``: Contact information of people whom can be reached
+- ``"Mode description"``: A description of the model.
+- ``"Intended uses & limitations"``: Intended use for the model, limitations and
+  potential biases. This section should also include risks of using models in
+  certain domains if relevant.
+- ``"How to Get Started with the Model"``: Code the user can run to load and use
+  the model.
+- ``"Model Card Authors"``: Authors of the model card. This section includes
+  authors of the model card
+- ``"Model Card Contact"``: Contact information of people whom can be reached
   out, in case of questions about the model or the model card.
-- ``"citation_bibtex"``: Bibtex style citations for the model or resources used
-  to train the model.
-- ``"eval_methods"``: Details about evaluation process of the model.
-- ``"eval_results"``: Evaluation results that are later parsed as a table by
-  :class:`skops.card.Card`.
+- ``"Citation"``: Bibtex style citations for the model or resources used to
+  train the model.
+- ``"Evaluation Results"``: Evaluation results that are later parsed as a table
+  by :class:`skops.card.Card`.
 
 
 The template also contains the following sections that are automatically
 generated by ``skops``.
 
-- ``"hyperparameter_table"``: Hyperparameters of the model.
-- ``"model_plot"``: A diagram of the model, most relevant in case the model is
+- ``"Hyperparameters"``: Hyperparameters of the model.
+- ``"Model Plot"``: A diagram of the model, most relevant in case the model is
   a complex scikit-learn :class:`~sklearn.pipeline.Pipeline`.
 
 Furthermore, it is possible to add plots and tables to the model card. To add
@@ -76,6 +74,32 @@ to the :meth:`.Card.add_plot` method. For tables, you can pass either
 dictionaries with the key being the header and the values being list of row
 entries, or a pandas ``DataFrame``; use the :meth:`.Card.add_table` method for
 this.
+
+To add content to an existing subsection, or create a new subsection, use a
+``"/"`` to indicate the subsection. E.g. let's assume you would like to add a
+subsection called ``"Figures"`` to the existing section ``"Model description"``,
+as well as adding some subsections with plots below that, you can call the
+:meth:`Card.add` method like this:
+
+.. code-block:: python
+
+    card.add(**{"Model description/Figures": "Here are some nice figures"})
+    card.add_plot(**{
+        "Model description/Figures/Confusion Matrix": "path-to-confusion-matrix.png",
+        "Model description/Figures/ROC": "path-to-roc.png",
+    })
+
+Furthermore, you can select existing sections (as well as their subsections)
+using :meth:`Card.select`, and you can delete sections using
+:meth:`Card.delete`:
+
+.. code-block:: python
+
+    section = card.select("Model description/Figures")
+    print(section.content)  # 'Here are some nice figures'
+    print(section.subsections)
+    card.delete("Model description/Figures/ROC")
+
 
 To see how you can use the API in ``skops`` to create a model card, please
 refer to :ref:`sphx_glr_auto_examples_plot_model_card.py`.
