+++ skops/card/_model_card.py
@@ -1,25 +1,52 @@
 from __future__ import annotations
 
-import copy
 import json
 import re
-import shutil
-import tempfile
-from dataclasses import dataclass
+import sys
+import textwrap
+from dataclasses import dataclass, field
 from pathlib import Path
 from reprlib import Repr
-from typing import Any, Optional, Union
+from typing import Any, Iterator, Sequence, Union
 
-from huggingface_hub import CardData, ModelCard
+from huggingface_hub import CardData
 from sklearn.utils import estimator_html_repr
 from tabulate import tabulate  # type: ignore
 
-import skops
+if sys.version_info >= (3, 8):
+    from typing import Protocol
+else:
+    from typing_extensions import Protocol
 
 # Repr attributes can be used to control the behavior of repr
 aRepr = Repr()
 aRepr.maxother = 79
 aRepr.maxstring = 79
+
+
+CONTENT_PLACEHOLDER = "[More Information Needed]"
+"""When there is a section but no content, show this"""
+
+DEFAULT_TEMPLATE = {
+    "Model description": CONTENT_PLACEHOLDER,
+    "Model description/Intended uses & limitations": CONTENT_PLACEHOLDER,
+    "Model description/Training Procedure": "",
+    "Model description/Training Procedure/Hyperparameters": CONTENT_PLACEHOLDER,
+    "Model description/Training Procedure/Model Plot": CONTENT_PLACEHOLDER,
+    "Model description/Evaluation Results": CONTENT_PLACEHOLDER,
+    "How to Get Started with the Model": CONTENT_PLACEHOLDER,
+    "Model Card Authors": (
+        f"This model card is written by following authors:\n\n{CONTENT_PLACEHOLDER}"
+    ),
+    "Model Card Contact": (
+        "You can contact the model card authors through following channels:\n"
+        f"{CONTENT_PLACEHOLDER}"
+    ),
+    "Citation": (
+        "Below you can find information related to citation.\n\n**BibTeX:**\n```\n"
+        f"{CONTENT_PLACEHOLDER}\n```"
+    ),
+}
 
 
 def wrap_as_details(text: str, folded: bool) -> str:
@@ -162,13 +189,106 @@ def metadata_from_config(config_path: Union[str, Path]) -> CardData:
     return card_data
 
 
+def split_subsection_names(key: str) -> list[str]:
+    r"""Split a string containing multiple sections into a list of strings for
+    each.
+
+    The separator is ``"/"``. To avoid splitting on ``"/"``, escape it using
+    ``"\\/"``.
+
+    Examples
+    --------
+    >>> split_subsection_names("Section A")
+    ['Section A']
+    >>> split_subsection_names("Section A/Section B/Section C")
+    ['Section A', 'Section B', 'Section C']
+    >>> split_subsection_names("A section containg \\/ a slash")
+    ['A section containg / a slash']
+    >>> split_subsection_names("Spaces are / stripped")
+    ['Spaces are', 'stripped']
+
+    Parameters
+    ----------
+    key : str
+        The section name consisting potentially of multiple subsections. It has
+        to be ensured beforhand that this is not an empty string.
+
+    Returns
+    -------
+    parts : list of str
+        The individual (sub)sections.
+
+    """
+    placeholder = "$%!?"  # arbitrary sting that never appears naturally
+    key = key.replace("\\/", placeholder)
+    parts = (part.strip() for part in key.split("/"))
+    return [part.replace(placeholder, "/") for part in parts]
+
+
+def _getting_started_code(
+    file_name: str, is_skops_format: bool = False, indent="    "
+) -> list[str]:
+    # get lines of code required to load the model
+    lines: list[str] = []
+    if is_skops_format:
+        lines += ["from skops.io import load"]
+    else:
+        lines += ["import joblib"]
+
+    lines += [
+        "import json",
+        "import pandas as pd",
+    ]
+    if is_skops_format:
+        lines += [
+            "from skops.io import load",
+            f'model = load("{file_name}")',
+        ]
+    else:  # pickle
+        lines += [f"model = joblib.load({file_name})"]
+
+    lines += [
+        'with open("config.json") as f:',
+        indent + "config = json.load(f)",
+        'model.predict(pd.DataFrame.from_dict(config["sklearn"]["example_input"]))',
+    ]
+    return lines
+
+
+@dataclass
+class Section:
+    """Building block of the model card.
+
+    The model card is represented internally as a dict with keys being strings
+    and values being Sections. The key is identical to the section title.
+
+    Additionally, the section may hold content in the form of strings (can be an
+    empty string) or a ``Formattable``, which is simply an object with a
+    ``format`` method that returns a string.
+
+    Finally, the section can contain subsections, which again are dicts of
+    string keys and section values (the dict can be empty). Therefore, the model
+    card representation forms a tree structure, making use of the fact that dict
+    order is preserved.
+
+    """
+
+    title: str
+    content: Formattable | str
+    subsections: dict[str, Section] = field(default_factory=dict)
+
+
+class Formattable(Protocol):
+    def format(self) -> str:
+        ...  # pragma: no cover
+
+
 class Card:
     """Model card class that will be used to generate model card.
 
     This class can be used to write information and plots to model card and save
     it. This class by default generates an interactive plot of the model and a
-    table of hyperparameters. The slots to be filled are defined in the markdown
-    template.
+    table of hyperparameters. Some sections are added by default.
 
     Parameters
     ----------
@@ -187,6 +307,9 @@ class Card:
         of the ``config.json`` file, which itself is created by
         :func:`skops.hub_utils.init`.
 
+    prefill: bool (default=True)
+        Whether to add default sections or not.
+
     Attributes
     ----------
     model: estimator object
@@ -195,13 +318,6 @@ class Card:
     metadata: CardData
         Metadata to be stored at the beginning of the saved model card, as
         metadata to be understood by the Hugging Face Hub.
-
-    Notes
-    -----
-    The contents of the sections of the template can be set using
-    :meth:`Card.add` method. Plots can be added to the model card using
-    :meth:`Card.add_plot`. The key you pass to :meth:`Card.add_plot` will be
-    used as the header of the plot.
 
     Examples
     --------
@@ -215,10 +331,10 @@ class Card:
     >>> from pathlib import Path
     >>> from sklearn.datasets import load_iris
     >>> from sklearn.linear_model import LogisticRegression
-    >>> from skops import card
+    >>> from skops.card import Card
     >>> X, y = load_iris(return_X_y=True)
     >>> model = LogisticRegression(solver="liblinear", random_state=0).fit(X, y)
-    >>> model_card = card.Card(model)
+    >>> model_card = Card(model)
     >>> model_card.metadata.license = "mit"
     >>> y_pred = model.predict(X)
     >>> model_card.add_metrics(**{
@@ -226,8 +342,11 @@ class Card:
     ...     "f1 score": f1_score(y, y_pred, average="micro"),
     ... })
     Card(
-      model=LogisticRegression(random_state=0, solver='liblinear'),
+      model=LogisticRegression(random_state=0, solver='liblinear')
       metadata.license=mit,
+      Model description/Training Procedure/... | | warm_start | False | </details>,
+      Model description/Training Procedure/...</pre></div></div></div></div></div>,
+      Model description/Evaluation Results=...ccuracy | 0.96 | | f1 score | 0.96 |,
     )
     >>> cm = confusion_matrix(y, y_pred,labels=model.classes_)
     >>> disp = ConfusionMatrixDisplay(
@@ -236,51 +355,297 @@ class Card:
     ... )
     >>> disp.plot()
     <sklearn.metrics._plot.confusion_matrix.ConfusionMatrixDisplay object at ...>
-    >>> disp.figure_.savefig("confusion_matrix.png")
+    >>> tmp_path = Path(tempfile.mkdtemp(prefix="skops-"))
+    >>> disp.figure_.savefig(tmp_path / "confusion_matrix.png")
     ...
-    >>> model_card.add_plot(confusion_matrix="confusion_matrix.png")
+    >>> model_card.add_plot(**{
+    ...     "Model description/Confusion Matrix": tmp_path / "confusion_matrix.png"
+    ... })
     Card(
-      model=LogisticRegression(random_state=0, solver='liblinear'),
+      model=LogisticRegression(random_state=0, solver='liblinear')
       metadata.license=mit,
-      confusion_matrix='confusion_matrix.png',
+      Model description/Training Procedure/... | | warm_start | False | </details>,
+      Model description/Training Procedure/...</pre></div></div></div></div></div>,
+      Model description/Evaluation Results=...ccuracy | 0.96 | | f1 score | 0.96 |,
+      Model description/Confusion Matrix=...confusion_matrix.png'),
     )
-    >>> with tempfile.TemporaryDirectory() as tmpdir:
-    ...     model_card.save((Path(tmpdir) / "README.md"))
+    >>> # add new content to the existing section "Model description"
+    >>> model_card.add(**{"Model description": "This is the best model"})
+    Card(
+      model=LogisticRegression(random_state=0, solver='liblinear')
+      metadata.license=mit,
+      Model description=This is the best model,
+      Model description/Training Procedure/... | | warm_start | False | </details>,
+      Model description/Training Procedure/...</pre></div></div></div></div></div>,
+      Model description/Evaluation Results=...ccuracy | 0.96 | | f1 score | 0.96 |,
+      Model description/Confusion Matrix=...confusion_matrix.png'),
+    )
+    >>> # add content to a new section
+    >>> model_card.add(**{"A new section": "Please rate my model"})
+    Card(
+      model=LogisticRegression(random_state=0, solver='liblinear')
+      metadata.license=mit,
+      Model description=This is the best model,
+      Model description/Training Procedure/... | | warm_start | False | </details>,
+      Model description/Training Procedure/...</pre></div></div></div></div></div>,
+      Model description/Evaluation Results=...ccuracy | 0.96 | | f1 score | 0.96 |,
+      Model description/Confusion Matrix=...confusion_matrix.png'),
+      A new section=Please rate my model,
+    )
+    >>> # add new subsection to an existing section by using "/"
+    >>> model_card.add(**{"Model description/Model name": "This model is called Bob"})
+    Card(
+      model=LogisticRegression(random_state=0, solver='liblinear')
+      metadata.license=mit,
+      Model description=This is the best model,
+      Model description/Training Procedure/... | | warm_start | False | </details>,
+      Model description/Training Procedure/...</pre></div></div></div></div></div>,
+      Model description/Evaluation Results=...ccuracy | 0.96 | | f1 score | 0.96 |,
+      Model description/Confusion Matrix=...confusion_matrix.png'),
+      Model description/Model name=This model is called Bob,
+      A new section=Please rate my model,
+    )
+    >>> # save the card to a README.md file
+    >>> model_card.save(tmp_path / "README.md")
     """
 
     def __init__(
         self,
-        model: Any,
+        model,
         model_diagram: bool = True,
-        metadata: Optional[CardData] = None,
+        metadata: CardData | None = None,
+        prefill: bool = True,
     ) -> None:
         self.model = model
         self.model_diagram = model_diagram
-        self._eval_results = {}  # type: ignore
-        self._template_sections: dict[str, str] = {}
-        self._extra_sections: list[tuple[str, Any]] = []
         self.metadata = metadata or CardData()
 
-    def add(self, **kwargs: str) -> "Card":
-        """Takes values to fill model card template.
+        self._data: dict[str, Section] = {}
+        self._metrics: dict[str, str | float | int] = {}
+        if prefill:
+            self._fill_default_sections()
+            # TODO: This is for parity with old model card but having an empty
+            # table by default is kinda pointless
+            self.add_metrics()
+            self._reset()
+
+    def _reset(self) -> None:
+        model_file = self.metadata.to_dict().get("model_file")
+        if model_file:
+            self._add_get_started_code(model_file)
+
+        self._add_model_section()
+        self._add_hyperparams()
+
+    def _fill_default_sections(self) -> None:
+        self.add(**DEFAULT_TEMPLATE)
+
+    def add(self, **kwargs: str | Formattable) -> "Card":
+        """Add new section(s) to the model card.
+
+        Add one or multiple sections to the model card. The section names are
+        taken from the keys and the contents are taken from the values.
+
+        To add to an existing section, use a ``"/"`` in the section name, e.g.:
+
+        ``card.add(**{"Existing section/New section": "content"})``.
+
+        If the parent section does not exist, it will be added automatically.
+
+        To add a section with ``"/"`` in its title (i.e. not inteded as a
+        subsection), escape the slash like so, ``"\\/"``, e.g.:
+
+        ``card.add(**{"A section with\\/a slash in the title": "content"})``.
+
+        If a section of the given name already exists, its content will be
+        overwritten.
 
         Parameters
         ----------
         **kwargs : dict
-            Parameters to be set for the model card. These parameters
-            need to be sections of the underlying `jinja` template used.
+            The keys of the dictionary serve as the section title and the values
+            as the section content. It's possible to add to existing sections.
 
         Returns
         -------
         self : object
             Card object.
+
         """
-        for section, value in kwargs.items():
-            self._template_sections[section] = value
+        for key, val in kwargs.items():
+            self._add_single(key, val)
         return self
 
-    def add_plot(self, folded=False, **kwargs: str) -> "Card":
+    def _select(
+        self, subsection_names: Sequence[str], create: bool = True
+    ) -> dict[str, Section]:
+        section = self._data
+        if not subsection_names:
+            return section
+
+        for subsection_name in subsection_names:
+            section_maybe = section.get(subsection_name)
+
+            # there are already subsections
+            if section_maybe is not None:
+                section = section_maybe.subsections
+                continue
+
+            if create:
+                # no subsection, create
+                entry = Section(title=subsection_name, content="")
+                section[subsection_name] = entry
+                section = entry.subsections
+            else:
+                raise KeyError(f"Section {subsection_name} does not exist")
+
+        return section
+
+    def select(self, key: str | Sequence[str]) -> Section:
+        """Select a section from the model card.
+
+        To select a subsection of an existing section, use a ``"/"`` in the
+        section name, e.g.:
+
+        ``card.select("Existing section/New section")``.
+
+        Alternatively, a list of strings can be passed:
+
+        ``card.select(["Existing section", "New section"])``.
+
+        Parameters
+        ----------
+        key : str or list of str
+            The name of the (sub)section to select. When selecting a subsection,
+            either use a ``"/"`` in the name to separate the parent and child
+            sections, or pass a list of strings.
+
+        Returns
+        -------
+        self : Section
+            A dataclass containing all information relevant to the selected
+            section. Those are the title, the content, and subsections (in a
+            dict).
+
+        Raises
+        ------
+        KeyError
+            If the given section name was not found, a ``KeyError`` is raised.
+
+        """
+        if not key:
+            msg = f"Section name cannot be empty but got '{key}'"
+            raise KeyError(msg)
+
+        if isinstance(key, str):
+            *subsection_names, leaf_node_name = split_subsection_names(key)
+        else:
+            *subsection_names, leaf_node_name = key
+
+        if not leaf_node_name:
+            msg = f"Section name cannot be empty but got '{key}'"
+            raise KeyError(msg)
+
+        parent_section = self._select(subsection_names, create=False)
+        return parent_section[leaf_node_name]
+
+    def delete(self, key: str | Sequence[str]) -> None:
+        """Delete a section from the model card.
+
+        To delete a subsection of an existing section, use a ``"/"`` in the
+        section name, e.g.:
+
+        ``card.delete("Existing section/New section")``.
+
+        Alternatively, a list of strings can be passed:
+
+        ``card.delete(["Existing section", "New section"])``.
+
+        Parameters
+        ----------
+        key : str or list of str
+            The name of the (sub)section to select. When selecting a subsection,
+            either use a ``"/"`` in the name to separate the parent and child
+            sections, or pass a list of strings.
+
+        Raises
+        ------
+        KeyError
+            If the given section name was not found, a ``KeyError`` is raised.
+
+        """
+        if not key:
+            msg = f"Section name cannot be empty but got '{key}'"
+            raise KeyError(msg)
+
+        if isinstance(key, str):
+            *subsection_names, leaf_node_name = split_subsection_names(key)
+        else:
+            *subsection_names, leaf_node_name = key
+
+        if not leaf_node_name:
+            msg = f"Section name cannot be empty but got '{key}'"
+            raise KeyError(msg)
+
+        parent_section = self._select(subsection_names, create=False)
+        del parent_section[leaf_node_name]
+
+    def _add_single(self, key: str, val: Formattable | str) -> None:
+        *subsection_names, leaf_node_name = split_subsection_names(key)
+        section = self._select(subsection_names)
+
+        if leaf_node_name in section:
+            # entry exists, only overwrite content
+            section[leaf_node_name].content = val
+        else:
+            # entry does not exist, create a new one
+            section[leaf_node_name] = Section(title=leaf_node_name, content=val)
+
+    def _add_model_section(self) -> None:
+        section_title = "Model description/Training Procedure/Model Plot"
+        default_content = "The model plot is below."
+
+        if not self.model_diagram:
+            self._add_single(section_title, default_content)
+            return
+
+        model_plot_div = re.sub(r"\n\s+", "", str(estimator_html_repr(self.model)))
+        if model_plot_div.count("sk-top-container") == 1:
+            model_plot_div = model_plot_div.replace(
+                "sk-top-container", 'sk-top-container" style="overflow: auto;'
+            )
+        content = f"{default_content}\n\n{model_plot_div}"
+        self._add_single(section_title, content)
+
+    def _add_hyperparams(self) -> None:
+        hyperparameter_dict = self.model.get_params(deep=True)
+        table = _clean_table(
+            tabulate(
+                list(hyperparameter_dict.items()),
+                headers=["Hyperparameter", "Value"],
+                tablefmt="github",
+            )
+        )
+        template = textwrap.dedent(
+            """        The model is trained with below hyperparameters.
+
+        <details>
+        <summary> Click to expand </summary>
+
+        {}
+
+        </details>"""
+        )
+        self._add_single(
+            "Model description/Training Procedure/Hyperparameters",
+            template.format(table),
+        )
+
+    def add_plot(self, *, folded=False, **kwargs: str) -> "Card":
         """Add plots to the model card.
+
+        The plot should be saved on the file system and the path passed as
+        value.
 
         Parameters
         ----------
@@ -291,22 +656,27 @@ class Card:
             large.
 
         **kwargs : dict
-            The arguments should be of the form `name=plot_path`, where `name`
-            is the name of the plot and `plot_path` is the path to the plot,
-            relative to the root of the project. The plots should have already
-            been saved under the project's folder.
+            The arguments should be of the form ``name=plot_path``, where
+            ``name`` is the name of the plot and section, and ``plot_path`` is
+            the path to the plot on the file system, relative to the root of the
+            project. The plots should have already been saved under the
+            project's folder.
 
         Returns
         -------
         self : object
             Card object.
+
         """
-        for plot_name, plot_path in kwargs.items():
+        for section_name, plot_path in kwargs.items():
+            plot_name = split_subsection_names(section_name)[-1]
             section = PlotSection(alt_text=plot_name, path=plot_path, folded=folded)
-            self._extra_sections.append((plot_name, section))
+            self._add_single(section_name, section)
         return self
 
-    def add_table(self, folded: bool = False, **kwargs: dict["str", list[Any]]) -> Card:
+    def add_table(
+        self, *, folded: bool = False, **kwargs: dict["str", list[Any]]
+    ) -> Card:
         """Add a table to the model card.
 
         Add a table to the model card. This can be especially useful when you
@@ -353,110 +723,156 @@ class Card:
         """
         for key, val in kwargs.items():
             section = TableSection(table=val, folded=folded)
-            self._extra_sections.append((key, section))
+            self._add_single(key, section)
         return self
 
-    def add_metrics(self, **kwargs: str) -> "Card":
+    def add_metrics(self, **kwargs: str | int | float) -> "Card":
         """Add metric values to the model card.
 
         Parameters
         ----------
         **kwargs : dict
-            A dictionary of the form `{metric name: metric value}`.
+            A dictionary of the form ``{metric name: metric value}``.
 
         Returns
         -------
         self : object
             Card object.
         """
-        for metric, value in kwargs.items():
-            self._eval_results[metric] = value
+        self._metrics.update(kwargs)
+        self._add_metrics(self._metrics)
         return self
 
-    def _generate_card(self) -> ModelCard:
-        """Generate the ModelCard object
-
-        Returns
-        -------
-        card : huggingface_hub.ModelCard
-            The final :class:`huggingface_hub.ModelCard` object with all
-            placeholders filled and all extra sections inserted.
-        """
-        root = skops.__path__
-
-        # add evaluation results
-
-        template_sections = copy.deepcopy(self._template_sections)
-
-        if self.metadata:
-            if self.metadata.to_dict().get("model_file"):
-                model_file = self.metadata.to_dict().get("model_file")
-                if model_file.endswith(".skops"):
-                    template_sections["get_started_code"] = (
-                        "from skops.io import load\nimport json\n"
-                        "import pandas as pd\n"
-                        f'model = load("{model_file}")\n'
-                        'with open("config.json") as f:\n   '
-                        " config ="
-                        " json.load(f)\n"
-                        'model.predict(pd.DataFrame.from_dict(config["sklearn"]["example_input"]))'
-                    )
-                else:
-                    template_sections["get_started_code"] = (
-                        "import joblib\nimport json\nimport pandas as pd\nmodel ="
-                        f' joblib.load({model_file})\nwith open("config.json") as'
-                        " f:\n   "
-                        " config ="
-                        " json.load(f)\n"
-                        'model.predict(pd.DataFrame.from_dict(config["sklearn"]["example_input"]))'
-                    )
-        if self.model_diagram is True:
-            model_plot_div = re.sub(r"\n\s+", "", str(estimator_html_repr(self.model)))
-            if model_plot_div.count("sk-top-container") == 1:
-                model_plot_div = model_plot_div.replace(
-                    "sk-top-container", 'sk-top-container" style="overflow: auto;'
-                )
-            model_plot: str | None = model_plot_div
-        else:
-            model_plot = None
-
-        template_sections["eval_results"] = tabulate(
-            list(self._eval_results.items()),
+    def _add_metrics(self, metrics: dict[str, str | float | int]) -> None:
+        table = tabulate(
+            list(metrics.items()),
             headers=["Metric", "Value"],
             tablefmt="github",
         )
+        template = textwrap.dedent(
+            """        You can find the details about evaluation process and the evaluation results.
 
-        # if template path is not given, use default
-        if template_sections.get("template_path") is None:
-            template_sections["template_path"] = str(
-                Path(root[0]) / "card" / "default_template.md"
-            )
 
-        # copying the template so that the original template is not touched/changed
-        # append plot_name if any plots are provided, at the end of the template
-        with tempfile.TemporaryDirectory() as tmpdirname:
-            shutil.copyfile(
-                template_sections["template_path"],
-                f"{tmpdirname}/temporary_template.md",
-            )
-            #  create a temporary template with the additional plots
-            template_sections["template_path"] = f"{tmpdirname}/temporary_template.md"
-            # add extra sections at the end of the template
-            with open(template_sections["template_path"], "a") as template:
-                if self._extra_sections:
-                    template.write("\n\n# Additional Content\n")
 
-                for key, val in self._extra_sections:
-                    formatted = val.format()
-                    template.write(f"\n## {key}\n\n{formatted}\n")
+        {}"""
+        )
+        self._add_single("Model description/Evaluation Results", template.format(table))
 
-            card = ModelCard.from_template(
-                card_data=self.metadata,
-                hyperparameter_table=self._extract_estimator_config(),
-                model_plot=model_plot,
-                **template_sections,
-            )
-        return card
+    def _generate_metadata(self, metadata: CardData) -> Iterator[str]:
+        """Yield metadata in yaml format"""
+        for key, val in metadata.to_dict().items() if metadata else {}:
+            yield aRepr.repr(f"metadata.{key}={val},").strip('"').strip("'")
+
+    def _generate_content(
+        self, data: dict[str, Section], depth: int = 1
+    ) -> Iterator[str]:
+        """Yield title and (formatted) contents"""
+        for val in data.values():
+            title = f"{depth * '#'} {val.title}"
+            yield title
+
+            if isinstance(val.content, str):
+                yield val.content
+            else:  # is a Formattable
+                yield val.content.format()
+
+            if val.subsections:
+                yield from self._generate_content(val.subsections, depth=depth + 1)
+
+    def _iterate_content(
+        self, data: dict[str, Section], parent_section: str = ""
+    ) -> Iterator[tuple[str, Formattable | str]]:
+        """Yield tuples of title and (non-formatted) content"""
+        for val in data.values():
+            if parent_section:
+                title = "/".join((parent_section, val.title))
+            else:
+                title = val.title
+
+            yield title, val.content
+
+            if val.subsections:
+                yield from self._iterate_content(val.subsections, parent_section=title)
+
+    @staticmethod
+    def _format_repr(text: str) -> str:
+        # Remove new lines, multiple spaces, quotation marks, and cap line length
+        text = text.replace("\n", " ")
+        text = re.sub(r"\s+", r" ", text)
+        return aRepr.repr(text).strip('"').strip("'")
+
+    def __str__(self) -> str:
+        return self.__repr__()
+
+    def __repr__(self) -> str:
+        # repr for the model
+        model = getattr(self, "model", None)
+        if model:
+            model_repr = self._format_repr(f"model={repr(model)}")
+        else:
+            model_repr = None
+
+        # repr for metadata
+        metadata_reprs = []
+        for key, val in self.metadata.to_dict().items() if self.metadata else {}:
+            if key == "widget":
+                metadata_reprs.append("metadata.widget={...},")
+                continue
+
+            metadata_reprs.append(self._format_repr(f"metadata.{key}={val},"))
+        metadata_repr = "\n".join(metadata_reprs)
+
+        # repr for contents
+        content_reprs = []
+        for title, content in self._iterate_content(self._data):
+            if not content:
+                continue
+            if isinstance(content, str) and content.rstrip("`").rstrip().endswith(
+                CONTENT_PLACEHOLDER
+            ):
+                # if content is just some default text, no need to show it
+                continue
+            content_reprs.append(self._format_repr(f"{title}={content},"))
+        content_repr = "\n".join(content_reprs)
+
+        # combine all parts
+        complete_repr = "Card(\n"
+        if model_repr:
+            complete_repr += textwrap.indent(model_repr, "  ") + "\n"
+        if metadata_reprs:
+            complete_repr += textwrap.indent(metadata_repr, "  ") + "\n"
+        if content_reprs:
+            complete_repr += textwrap.indent(content_repr, "  ") + "\n"
+        complete_repr += ")"
+        return complete_repr
+
+    def _add_get_started_code(self, file_name: str, indent: str = "    ") -> None:
+        is_skops_format = file_name.endswith(".skops")  # else, assume pickle
+        lines = _getting_started_code(
+            file_name, is_skops_format=is_skops_format, indent=indent
+        )
+        lines = ["```python"] + lines + ["```"]
+
+        template = textwrap.dedent(
+            """        Use the code below to get started with the model.
+
+        {}
+        """
+        )
+        self._add_single(
+            "How to Get Started with the Model", template.format("\n".join(lines))
+        )
+
+    def _generate_card(self) -> Iterator[str]:
+        if self.metadata.to_dict():
+            yield f"---\n{self.metadata.to_yaml()}\n---"
+
+        for line in self._generate_content(self._data):
+            if line:
+                yield "\n" + line
+
+        # add an empty line add the end
+        yield ""
 
     def save(self, path: str | Path) -> None:
         """Save the model card.
@@ -474,91 +890,16 @@ class Card:
         The keys in model card metadata can be seen `here
         <https://huggingface.co/docs/hub/models-cards#model-card-metadata>`__.
         """
-        card = self._generate_card()
-        card.save(path)
+        with open(path, "w", encoding="utf-8") as f:
+            f.write("\n".join(self._generate_card()))
 
     def render(self) -> str:
         """Render the final model card as a string.
 
         Returns
         -------
-        card : str
+        result : str
             The rendered model card with all placeholders filled and all extra
             sections inserted.
         """
-        card = self._generate_card()
-        return str(card)
-
-    def _extract_estimator_config(self) -> str:
-        """Extracts estimator hyperparameters and renders them into a vertical table.
-
-        Returns
-        -------
-        str:
-            Markdown table of hyperparameters.
-        """
-        hyperparameter_dict = self.model.get_params(deep=True)
-        return _clean_table(
-            tabulate(
-                list(hyperparameter_dict.items()),
-                headers=["Hyperparameter", "Value"],
-                tablefmt="github",
-            )
-        )
-
-    @staticmethod
-    def _strip_blank(text) -> str:
-        # remove new lines and multiple spaces
-        text = text.replace("\n", " ")
-        text = re.sub(r"\s+", r" ", text)
-        return text
-
-    def __str__(self) -> str:
-        return self.__repr__()
-
-    def __repr__(self) -> str:
-        # create repr for model
-        model = getattr(self, "model", None)
-        if model:
-            model_str = self._strip_blank(repr(model))
-            model_repr = aRepr.repr(f"  model={model_str},").strip('"').strip("'")
-        else:
-            model_repr = None
-
-        # metadata
-        metadata_reprs = []
-        for key, val in self.metadata.to_dict().items() if self.metadata else {}:
-            if key == "widget":
-                metadata_reprs.append("  metadata.widget={...},")
-                continue
-
-            metadata_reprs.append(
-                aRepr.repr(f"  metadata.{key}={val},").strip('"').strip("'")
-            )
-        metadata_repr = "\n".join(metadata_reprs)
-
-        # normal sections
-        template_reprs = []
-        for key, val in self._template_sections.items():
-            val = self._strip_blank(repr(val))
-            template_reprs.append(aRepr.repr(f"  {key}={val},").strip('"').strip("'"))
-        template_repr = "\n".join(template_reprs)
-
-        # figures
-        figure_reprs = []
-        for key, val in self._extra_sections:
-            val = self._strip_blank(repr(val))
-            figure_reprs.append(aRepr.repr(f"  {key}={val},").strip('"').strip("'"))
-        figure_repr = "\n".join(figure_reprs)
-
-        complete_repr = "Card(\n"
-        if model_repr:
-            complete_repr += model_repr + "\n"
-        if metadata_reprs:
-            complete_repr += metadata_repr + "\n"
-        if template_repr:
-            complete_repr += template_repr + "\n"
-        if figure_repr:
-            complete_repr += figure_repr + "\n"
-        complete_repr += ")"
-        return complete_repr
+        return "\n".join(self._generate_card())
